fix doc header/footer table section meta mis-use of :header/:footer keys
if a doc's header and/or footer is comprised of a table, a pdf-table
must be used. when evaluating the pdf-table element, the meta map passed
to it will be the same as the doc's meta map. as a result, the original
:header/:footer value would be passed to pdf-table which would cause
problems (completely wrong data for what pdf-table expects)